feat(*)!: add Message Events support
This feature introduces Message event support, with a large set of message subtypes.
Message attachements are not yet supported but they will in a near future.
You can find the list of supported event in a new documentation folder at the root of the repository.

Other noticeable changes:

* Simple sample has been updated to support this new type of event
* Usain.Slack models documentation has been largely enriched
* Timestamp equality operations have been fixed
* Huge catch up on the coverage (95% global)
* Third party dependencies have been upgraded

BREAKING CHANGES: Usain.Slack

A lot of types and type properties have been added, renamed or removed.
These changes will introduce a lot of minor breaking changes while staying relatively easy to fix.
This was a necessity to clean and stabilize the model API.